==> 1.music_data_manipulation.ipynb <==
import numpy as np
import pandas as pd
import os

folder = os.path.join(os.path.abspath(".."), "interest")

artist = pd.read_csv(os.path.join(folder, "data_by_artist.csv"))
year = pd.read_csv(os.path.join(folder, "data_by_year.csv"))
music = pd.read_csv(os.path.join(folder, "full_music_data.csv"))
influence = pd.read_csv(os.path.join(folder, "influence_data.csv"))

#### arrange music data

# preprocessing
# 修正数据类型
music["music_id"] = range(len(music))
music["release_date"] = music["release_date"].astype("datetime64")
music = music.sort_values(by = 'release_date', ascending = True)
music = music.reset_index(drop = True)

# 做pre和写report没用
# 因为原数据当中没有genre的信息，influence里面才有每个musician的genre，所以需要匹配一下每一首音乐对应的genre
# mapping music with genre
lookup1 = influence.loc[:, ["influencer_id", "influencer_main_genre"]]
lookup1.columns = ["artists_id", "genre"]

lookup2 = influence.loc[:, ["follower_id", "follower_main_genre"]]
lookup2.columns = ["artists_id", "genre"]

lookup = pd.concat([lookup1, lookup2])
lookup = lookup.groupby("artists_id").first().reset_index()

lookup_dict = lookup.set_index('artists_id')['genre'].to_dict()
# lookup_dict

lookup2 = influence.loc[:, ["follower_id", "follower_main_genre"]]
lookup2.columns = ["artists_id", "genre"]

lookup = pd.concat([lookup1, lookup2])
lookup = lookup.groupby("artists_id").first().reset_index()

lookup_dict = lookup.set_index('artists_id')['genre'].to_dict()

music_artists_lst = music["artists_id"].str.extract(r"\[(\d+(?:,\s*\d+)*)\]")[0].str.split(", ")

def artists_to_genre(lst):
    lst = [lookup_dict[int(artist_id)] if int(artist_id) in lookup_dict.keys() else "Unknown" for artist_id in lst]
    return max(set(lst), key = lst.count)

music["genre"] = music_artists_lst.apply(lambda x:artists_to_genre(x))

# delete niche groups
music = music.loc[music["genre"] != "Unknown", :]

# 去掉43年以前的音乐（因为数据太少）
print(music["year"].value_counts().tail())
music = music.loc[music["year"] > 1943, :]

music = music.reset_index(drop = True)

#### group by niche categories
#### 按照correlation of coefficient来合并genre
##### 最终，将"Comedy/Spoken", "Children's"去掉，'Avant-Garde',  'Classical',  'Jazz',  'Stage & Screen',  'Folk',  'Easy Listening',  'New Age',  'Vocal',  'International'合并为niche1，"Religious", "Electronic", "Country", "Latin", "New Age"合并为genre2

music.columns

feature_columns = ['danceability', 'energy', 'valence',
       'tempo', 'loudness', 'mode', 'key', 'acousticness', 'instrumentalness',
       'liveness', 'speechiness', 'explicit', 'duration_ms', 'popularity']

from sklearn.preprocessing import StandardScaler

tmp_scaler = StandardScaler()
music_tmp = tmp_scaler.fit_transform(music[feature_columns])
music_tmp = pd.DataFrame(music_tmp)

music_tmp.columns = feature_columns
music_tmp["music_id"] = music["music_id"]
music_tmp["genre"] = music["genre"]


genre_attributes = music_tmp.groupby(["genre"])[feature_columns].mean().T

genre_corr = genre_attributes.corr()

music["genre"].value_counts()

genre_corr.loc["Children's", :].sort_values(ascending = False).head()

genre_corr.loc["Avant-Garde", :].sort_values(ascending = False).head(10)

niche_genre1 = list(genre_corr.loc["Avant-Garde", :].sort_values(ascending = False).head(9).index)

niche_genre1

music["genre"] = music["genre"].apply(lambda x:"niche1" if x in niche_genre1 else x)

tmp_scaler = StandardScaler()
music_tmp = tmp_scaler.fit_transform(music[feature_columns])
music_tmp = pd.DataFrame(music_tmp)

music_tmp.columns = feature_columns
music_tmp["music_id"] = music["music_id"]
music_tmp["genre"] = music["genre"]

genre_attributes = music_tmp.groupby(["genre"])[feature_columns].mean().T

genre_corr = genre_attributes.corr()

music["genre"].value_counts()

genre_corr.loc["Religious", :].sort_values(ascending = False)

niche_genre2 = ["Religious", "Electronic", "Country", "Latin", "New Age"]

music["genre"] = music["genre"].apply(lambda x:"niche2" if x in niche_genre2 else x)

tmp_scaler = StandardScaler()
music_tmp = tmp_scaler.fit_transform(music[feature_columns])
music_tmp = pd.DataFrame(music_tmp)

music_tmp.columns = feature_columns
music_tmp["music_id"] = music["music_id"]
music_tmp["genre"] = music["genre"]

genre_attributes = music_tmp.groupby(["genre"])[feature_columns].mean().T

genre_corr = genre_attributes.corr()

music["genre"].value_counts()

genre_corr.loc["Reggae", :].sort_values(ascending = False)

music.loc[music["genre"] == "Reggae", "genre"] = "Pop/Rock"

tmp_scaler = StandardScaler()
music_tmp = tmp_scaler.fit_transform(music[feature_columns])
music_tmp = pd.DataFrame(music_tmp)

music_tmp.columns = feature_columns
music_tmp["music_id"] = music["music_id"]
music_tmp["genre"] = music["genre"]


genre_attributes = music_tmp.groupby(["genre"])[feature_columns].mean().T

genre_corr = genre_attributes.corr()

music["genre"].value_counts()

music = music.loc[~music["genre"].isin(["Comedy/Spoken", "Children's"]), :]
music = music.reset_index(drop = True)

music["genre"].value_counts()

#### add # of musician's information，增加每首歌有几位音乐家创作的feature

music_artists_lst = music["artists_id"].str.extract(r"\[(\d+(?:,\s*\d+)*)\]")[0].str.split(", ")
music_artists_lst

music["musician_cnt"] = music_artists_lst.apply(lambda x:len(x))

#### arrange the data

music.head()

music.columns

music = music.drop(columns = ['artist_names', 'year', "song_title (censored)"])
music = music.reset_index(drop = True)

#### Standardize music features，更改字符，通过方差均值进行标准化

music.columns

music["genre"].unique()

category_replace = {'Pop/Rock': "pop_rock",
                   "R&B;": "r_b"}

music["genre"] = music["genre"].replace(category_replace)

# add dummy variables
genre_dummy = pd.get_dummies(music_feature["genre"], drop_first = False)
music = pd.concat([music, genre_dummy], axis = 1)

music.columns

feature_columns = ['danceability', 'energy', 'valence', 'tempo', 'loudness',
       'mode', 'key', 'acousticness', 'instrumentalness', 'liveness',
       'speechiness', 'explicit', 'duration_ms', 'popularity', 'musician_cnt', 
                  'niche1', 'niche2', 'pop_rock', 'r_b']

scalar = StandardScaler()
music = music.reset_index(drop = True)
music_feature = scalar.fit_transform(music[feature_columns])

music_feature = pd.DataFrame(music_feature)
music_feature.columns = feature_columns

music_feature["artists_id"] = music["artists_id"]
music_feature["release_date"] = music["release_date"]
music_feature["music_id"] = music["music_id"]
music_feature["genre"] = music["genre"]

music_feature = music_feature.sort_values(by = ["release_date", "music_id"], ascending = True)
music_feature = music_feature.reset_index(drop = True)

music_feature.head()

music_feature.to_csv(r"rearrange_data/music_features.csv", index = False)
