==> 2.musician_sequential_input_gen.ipynb <==
import numpy as np
import pandas as pd
import os

folder = os.path.join(os.path.abspath(".."), "interest")

music = pd.read_csv(r"rearrange_data/music_features.csv") # 这个data是已经整理好的music，data，不再是最初的music feature了
influence = pd.read_csv(os.path.join(folder, "influence_data.csv"))

#### extract musician and music list

# 提取每个音乐对应的音乐家的id信息
music_artists_lst = music["artists_id"].str.extract(r"\[(\d+(?:,\s*\d+)*)\]")[0].str.split(", ")
artists_lst = music_artists_lst.values.tolist()
artists_lst = [item for sublist in artists_lst for item in sublist]
artists_lst = list(set(artists_lst)) # 找到所有数据库出现的音乐家

# 生成（key = musician_id, val = [music_id]）的dictionary
musician_creation = {}
for idx, artists_id in enumerate(music_artists_lst):
    music_id = music.iloc[idx, :]["music_id"]
    for artist in artists_id:
        if artist not in musician_creation.keys():
            musician_creation[artist] = [music_id]
        else:
            musician_creation[artist].append(music_id)

len(musician_creation) #一共有5000为音乐家

#### add year info

music["release_date"] = music["release_date"].astype("datetime64")
music["year"] = music["release_date"].dt.year

#### get the genre information

music.columns

feature_columns = ['danceability', 'energy', 'valence', 'tempo', 'loudness', 'mode', 'key',
       'acousticness', 'instrumentalness', 'liveness', 'speechiness',
       'explicit', 'duration_ms', 'popularity', 'musician_cnt', 'niche1',
       'niche2', 'pop_rock', 'r_b']

genre = music.groupby(["year", "genre"])[feature_columns].mean().reset_index() # 生成每个genre的music feature方便之后对于缺失值填充

genre = genre.sort_values(by = ["genre", "year"], ascending = True)
genre = genre.reset_index(drop = True)

genre_development_diminish = genre.copy() # 额外复制一个genre yearly development的df，方便填补缺失值
genre_development_diminish[feature_columns] = genre_development_diminish[feature_columns] * 0.8

genre_development_diminish.head()

# 用于根据musician id生成input data
def niche_time_gen(artist_id):
    # 找到该musician创作过的所有音乐
    music_id = musician_creation[artist_id]
    artist_pieces = music.loc[music["music_id"].isin(music_id), :] # related rows in music
    max_idx = len(artist_pieces) # how many rows in total
    if max_idx > 20:
        # 如果超过20首，随机取连续的20首音乐的feature作为input data
        selection = np.random.choice(range(0, max_idx - 20 + 1), 1, replace = False)[0] # randomly select a part of the music feature
        return artist_pieces.iloc[selection:(selection + 20), :][feature_columns], False
    else:
        # 如果不足20，则除了它创作过的音乐外，找到它活跃的最早年份，取出这一年以前的它所属genre的feature作为信息补齐够20条
        start_year = min(artist_pieces["year"]) # find the first year the musician start to make songs
        musician_genre = (artist_pieces["genre"].values[0]) # find the musician's genre
        related_genre_development = genre.loc[(genre["genre"] == musician_genre) & (genre["year"] < start_year), :] # extract the genre development info before the year
        if len(related_genre_development) == 0:
            related_genre_development = genre.loc[(genre["genre"] == musician_genre), :].head(1) # if too early, use the first year existed in the database
        miss = 20 - max_idx # how many rows to be filled
        related_genre_development = related_genre_development.tail(miss) # use the year before the oldest records of that genre to fill blanks
        
        empty_flag = False
        if len(related_genre_development) < miss: # if the genre information still not enough
            # 但是如果它所属genre有数据的年份也不够，则用该genre最早的一年的feature，按照每年80%的递减率，用最早一年进行递减补全
            empty_flag = True
            miss = miss - len(related_genre_development) # how many rows left
            baseline_idx = related_genre_development.index[0] # find the earliest year to fill it as the first t = -1 records
            baseline = genre_development_diminish.loc[baseline_idx, :][feature_columns].copy()
            empty = np.zeros((miss, 19))
            diminish_rate = 0.8
            for i in range(empty.shape[0] - 1, -1, -1):
                baseline *= diminish_rate
                empty[i, :] = baseline.copy()
            empty = pd.DataFrame(empty)
            empty.columns = feature_columns
                
        layer1 = (np.zeros((0, 19)) if not empty_flag else empty)
        layer2 = related_genre_development[feature_columns]
        layer3 = artist_pieces.loc[:, feature_columns]
        res = np.concatenate((layer1, layer2, layer3), axis = 0)
        return res, True

#### find influence relation

# 找到明确有influence的所有pair
influence_zip = influence.apply(lambda x: [str(x['influencer_id']), str(x['follower_id'])], axis=1).tolist()

len(artists_lst) ** 2

len(influence_zip) * 2

#### 先生成所有target variable为0的input data，一共生成5w组

# 从所有的musician当中随机生成pair
sample_lst = np.random.choice(artists_lst, (50000, 2), replace = True).tolist()

influencer_feature0 = []
follower_feature0 = []
niche_cnt = 0
cnt = 0
# 先生成所有target variable为0的input data，一共生成5w组
for pair in sample_lst:
    influencer_id = str(pair[0])
    follower_id = str(pair[1])
    if influencer_id in artists_lst and follower_id in artists_lst:
        if (influencer_id != follower_id) and (pair not in influence_zip):
            try:
                influencer_input, flag1 = niche_time_gen(influencer_id)
            except Exception as e:
                print(influencer_id, e)
                break
            try:
                follow_input, flag2 = niche_time_gen(follower_id)
            except Exception as e:
                print(follower_id, e)
                break    
            influencer_feature0.append(influencer_input)
            follower_feature0.append(follow_input)
            if flag1 or flag2:
                niche_cnt += 1
    if cnt % 100 == 0:
        print(cnt, influencer_id, follower_id)
    cnt += 1

import pickle

# 转换成numpy格式并保存
influencer_feature0 = np.array(influencer_feature0)
follower_feature0 = np.array(follower_feature0)

musician_sequential_0 = [influencer_feature0, follower_feature0]
with open(r'variables/musician_sequential_0.pkl', 'wb') as save_data:
    pickle.dump(musician_sequential_0, save_data) 

##### 生成所有target variable为1的input data

influencer_feature1 = []
follower_feature1 = []
niche_cnt = 0
cnt = 0
for pair in influence_zip:
    influencer_id = str(pair[0])
    follower_id = str(pair[1])
    if influencer_id in artists_lst and follower_id in artists_lst:
        try:
            influencer_input, flag1 = niche_time_gen(influencer_id)
        except Exception as e:
            print("influencer", influencer_id, e)
            break
        try:
            follow_input, flag2 = niche_time_gen(follower_id)
        except Exception as e:
            print("follower", follower_id, e)
            break    
        influencer_feature1.append(influencer_input)
        follower_feature1.append(follow_input)
        if flag1 or flag2:
            niche_cnt += 1
    if cnt % 100 == 0:
        print(cnt, influencer_id, follower_id)
    cnt += 1

influencer_feature1 = np.array(influencer_feature1)
follower_feature1 = np.array(follower_feature1)

musician_sequential_1 = [influencer_feature1, follower_feature1]
with open(r'variables/musician_sequential_1.pkl', 'wb') as save_data:
    pickle.dump(musician_sequential_1, save_data)
