==> 3.musician_influence.ipynb <==
import pickle
import os
import numpy as np

import warnings
warnings.filterwarnings("ignore")

folder = os.path.abspath(r'./variables')

filename = open(folder + r'/musician_sequential_0.pkl', 'rb')
sample0 = pickle.load(filename)

filename = open(folder + r'/musician_sequential_1.pkl', 'rb')
sample1 = pickle.load(filename)

influencer_feature = np.concatenate([sample0[0], sample1[0]])
follower_feature = np.concatenate([sample0[1], sample1[1]])

with_influence = len(sample1[0])
without_influence = len(sample0[0])

target = np.array([1 for _ in range(with_influence)] + [0 for _ in range(without_influence)])

idx = np.array(range(with_influence + without_influence))
np.random.shuffle(idx)

influencer_feature = influencer_feature[idx, :, :]
follower_feature = follower_feature[idx, :, :]
target = target[idx]

target.shape

# 按照7: 1.5: 1.5划分训练集、验证集、测试集
train_influencer = influencer_feature[:63801, :, :]
train_follower = follower_feature[:63801, :, :]
train_y = target[:63801]

val_influencer =  influencer_feature[63801:77473, :, :]
val_follower = follower_feature[63801:77473, :, :]
val_y = target[63801:77473]

test_influencer =  influencer_feature[77473:, :, :]
test_follower = follower_feature[77473:, :, :]
test_y = target[77473:]

from keras import layers
import keras
from keras import Input
from keras.models import Sequential, Model

monitor = [
    keras.callbacks.EarlyStopping(
        monitor = 'acc',
        patience = 3,
    ),
    keras.callbacks.ModelCheckpoint(
        filepath = 'influence_musicians.h5',
        monitor = 'val_loss',
        save_best_only = True,
    )
]

# RNN 结构
influencer_input = Input(shape = (20, 33), dtype = "float64", name = "influencer")
memory_influencer = layers.LSTM(512, input_shape = (20, 33), dropout = 0.3)(influencer_input)
memory_influencer = layers.Dense(512, activation = "relu")(memory_influencer)
memory_influencer = layers.Dropout(0.5)(memory_influencer)

follower_input = Input(shape = (20, 33), dtype = "float64", name = "follower") 
memory_follower = layers.LSTM(256, input_shape = (20, 33), dropout = 0.1)(follower_input)
memory_follower = layers.Dense(128, activation = "relu")(memory_follower)
memory_influencer = layers.Dropout(0.5)(memory_influencer)

concatenated = layers.concatenate([memory_influencer, memory_follower], axis = -1)

existance = layers.Dense(256, activation = "relu")(concatenated)
existance = layers.Dense(128, activation = "relu")(existance)
existance = layers.Dense(64, activation = "relu")(existance)
existance = layers.Dense(1, activation = "sigmoid")(existance)

model = Model([influencer_input, follower_input], existance)
optimizer = keras.optimizers.RMSprop(lr = 0.01)
model.compile(loss='binary_crossentropy', optimizer = optimizer, metrics = ["acc"])

history = model.fit([train_influencer, train_follower], train_y, 
          batch_size = 256, epochs = 100, 
          validation_data = ([val_influencer, val_follower], val_y), callbacks = monitor)

max(history.history["val_acc"])

from keras.models import load_model

influence_musicians = load_model("influence_musicians8570.h5")

influence_musicians.summary()

# 利用验证集找到benchmark来判断variable是0还是1

val_prob = influence_musicians.predict([val_influencer, val_follower])

hash_table = {}

from sklearn.metrics import accuracy_score

for bar in np.linspace(0.35, 0.45, 1000):
    val_pred = np.array([1 if val_prob[i] >= bar else 0 for i in range(len(val_prob))])
    accuracy = accuracy_score(val_y, val_pred)
    hash_table[bar] = accuracy

list(hash_table.keys())[np.argmax(np.array(list(hash_table.values())))]

bar = 0.40896

test_prob = influence_musicians.predict([test_influencer, test_follower])

test_pred = np.array([1 if test_prob[i] >= bar else 0 for i in range(len(test_prob))])

accuracy = accuracy_score(test_y, test_pred)
accuracy

from sklearn.metrics import roc_curve, auc

fpr, tpr, _  = roc_curve(test_y, test_pred)
auc(fpr, tpr)
